requirements.txt
aiohttp==3.7.4
pymongo==3.12.3
pymodm==0.4.3
google-api-python-client==1.6.5
slackclient==1.1.3
requests>=2.20.0
websocket-client==0.48.0
slacksocket==1.0.0
slack-sdk==3.4.0
beautifulsoup4==4.7.1
soupsieve==1.9
dnspython==1.16.0
dnslib==0.9.14
PyGithub==1.57